requirements.txt
web3~=7.3.0
mysql-connector-python~=9.0.0
SQLAlchemy~=2.0.32
requests~=2.32.3
fastapi~=0.63.0
python-multipart~=0.0.12
uvicorn~=0.13.3
starlette~=0.40.0
gTTS~=2.5.3
moviepy~=1.0.3
pydub~=0.25.1